chore: Unit testing without TestBed
- remove duplicated specs with TestBed
