+++ src/app/store/app.effects.spec.ts
@@ -1,102 +1,68 @@
-import { TestBed } from '@angular/core/testing';
-import { provideMockActions } from '@ngrx/effects/testing';
 import { ROUTER_NAVIGATION } from '@ngrx/router-store';
-import { MockStore, provideMockStore } from '@ngrx/store/testing';
-import { Observable, of } from 'rxjs';
+import { Action, ActionsSubject } from '@ngrx/store';
+import { MockStore } from '@ngrx/store/testing';
+import { of } from 'rxjs';
+import { Mock, mockActions, mockService, mockStore } from 'src/test-utils';
 import { UserService } from '../data/users.service';
 import * as AppActions from './app.actions';
 import { AppEffects } from './app.effects';
 import { AppState } from './app.reducer';
+import { selectAppUsers } from './app.selectors';
 
-const initialState = {
-  users: null,
-  settings: { enrolled: false },
-  signedIn: true,
-  accordionOpen: false,
-  error: null,
-};
-
-const mockUsers = [
-  {
-    id: 1,
-    name: 'Leanne Graham',
-    username: 'Bret',
-    email: 'Sincere@april.biz',
-    address: {
-      street: 'Kulas Light',
-      suite: 'Apt. 556',
-      city: 'Gwenborough',
-      zipcode: '92998-3874',
-      geo: {
-        lat: '-37.3159',
-        lng: '81.1496',
-      },
-    },
-    phone: '1-770-736-8031 x56442',
-    website: 'hildegard.org',
-    company: {
-      name: 'Romaguera-Crona',
-      catchPhrase: 'Multi-layered client-server neural-net',
-      bs: 'harness real-time e-markets',
-    },
-  },
-];
-
-class MockUserService {
-  fetchUsers() {
-    return of(mockUsers);
-  }
-}
-
-describe('AppEffects', () => {
-  let actions$: Observable<any>;
+describe('AppEffects 2', () => {
+  let actions: ActionsSubject;
   let effects: AppEffects;
   let store: MockStore<AppState>;
-  let httpService: UserService;
+  let userService: Mock<UserService>;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [
-        AppEffects,
-        provideMockActions(() => actions$),
-        provideMockStore({ initialState }),
-        { provide: UserService, useClass: MockUserService },
-      ],
-    });
+    store = mockStore();
+    actions = mockActions();
+    userService = mockService(UserService);
 
-    effects = TestBed.inject(AppEffects);
-    store = TestBed.inject(MockStore);
-    httpService = TestBed.inject(UserService);
+    effects = new AppEffects(actions, userService, store);
   });
 
   it('should be created', () => {
     expect(effects).toBeTruthy();
   });
 
-  // Testing API interaction
   describe('onFetchUsers$', () => {
     it('should fire if users is null', (done) => {
-      const spy = spyOn(httpService, 'fetchUsers').and.callThrough();
-      actions$ = of(AppActions.fetchUsers);
+      // given
+      const mockUsers = [{ id: 1, name: 'Leanne Graham' }];
+
+      userService.fetchUsers.and.returnValue(of(mockUsers));
+      store.overrideSelector(selectAppUsers, null);
+
+      // when
+      actions.next(AppActions.fetchUsers());
+
+      // then
       effects.onFetchUsers$.subscribe((res) => {
         expect(res).toEqual(AppActions.fetchUsersSuccess({ data: mockUsers }));
-        expect(spy).toHaveBeenCalledTimes(1);
+        expect(userService.fetchUsers).toHaveBeenCalled();
         done();
       });
     });
   });
 
-  // Testing interaction with Router-Store
   describe('onSignOut$', () => {
     it('should dispatch userSignOut if route matches', (done) => {
-      actions$ = of({
+      // given
+      const signoutRouterAction: Action = {
         type: ROUTER_NAVIGATION,
         payload: {
           routerState: {
             url: 'signout',
           },
         },
-      });
+      } as Action;
+
+      // when
+      actions.next(signoutRouterAction);
+
+      // then
       effects.onSignOut$.subscribe((action) => {
         expect(action).toEqual(AppActions.userSignOut());
         done();
@@ -104,12 +70,15 @@ describe('AppEffects', () => {
     });
   });
 
-  // Testing non-dispatching side-effect
   describe('onEnroll$', () => {
     it('should set item in local storage', (done) => {
+      // given
       const spy = spyOn(localStorage, 'setItem').and.callFake(() => true);
 
-      actions$ = of(AppActions.userEnroll());
+      // when
+      actions.next(AppActions.userEnroll());
+
+      // then
       effects.onEnroll$.subscribe((action) => {
         expect(action).toEqual(AppActions.userEnroll());
         expect(spy).toHaveBeenCalled();
